Swap again arguments
When a default parameter is used, it doesnt count to the total arity of
the function, e.g. f(a,b,c=1).length === 2, not 3.
When composing functions the initial approach to have the data structure
to be acted upon to be placed before the last argument and have the
default parameter autofilled doesnt work(limitations of JS). Now when
currying a function, the default parameter has to be explicitly defined
and also the total number of arguments of the curried function has to be
explicitly stated.
THIS IS ONLY THE CASE FOR DEFAULT PARAMETERS IN THE
FUNCTION DECLARATION.